fix: (core) fix after review
