+++ libs/core/src/lib/slider/constants.ts
@@ -1,0 +1,1 @@
+export const MIN_DISTANCE_BETWEEN_TICKS = 8;
