+++ libs/core/src/lib/slider/slider.component.html
@@ -8,12 +8,7 @@
                 [attr.tabindex]="disabled ? -1 : 0"
                 role="slider"
                 aria-label="slider"
-                [fdPopover]="singleSliderPopover"
-                [triggers]="[]"
-                [closeOnOutsideClick]="false"
-                [appendTo]="handle"
-                [disabled]="!tooltipMode"
-                placement="top"
+                [fdPopoverTrigger]="singleSliderPopover"
                 [attr.aria-valuemin]="min"
                 [attr.aria-valuemax]="max"
                 [attr.aria-valuenow]="getValuenow(_position, _sliderValueTargets.SINGLE_SLIDER)"
@@ -43,12 +38,7 @@
                 [attr.tabindex]="disabled ? -1 : 0"
                 role="slider"
                 aria-label="slider"
-                [fdPopover]="rangeSliderPopover1"
-                [triggers]="[]"
-                [closeOnOutsideClick]="false"
-                [appendTo]="rangeHandle1"
-                [disabled]="!tooltipMode"
-                placement="top"
+                [fdPopoverTrigger]="rangeSliderPopover1"
                 [attr.aria-valuemin]="min"
                 [attr.aria-valuemax]="max"
                 [attr.aria-valuenow]="getValuenow(_handle1Value, _sliderValueTargets.RANGE_SLIDER1)"
@@ -64,12 +54,7 @@
                 [attr.tabindex]="disabled ? -1 : 0"
                 role="slider"
                 aria-label="slider"
-                [fdPopover]="rangeSliderPopover2"
-                [triggers]="[]"
-                [closeOnOutsideClick]="false"
-                [appendTo]="rangeHandle2"
-                [disabled]="!tooltipMode"
-                placement="top"
+                [fdPopoverTrigger]="rangeSliderPopover2"
                 [attr.aria-valuemin]="min"
                 [attr.aria-valuemax]="max"
                 [attr.aria-valuenow]="getValuenow(_handle2Value, _sliderValueTargets.RANGE_SLIDER2)"
@@ -114,29 +99,53 @@
     </div>
 </div>
 
-<ng-template #singleSliderPopover>
+<fd-popover
+    #singleSliderPopover
+    *ngIf="!_isRange"
+    [triggers]="[]"
+    [closeOnOutsideClick]="false"
+    [appendTo]="handle"
+    [disabled]="!tooltipMode"
+    placement="top"
+>
     <ng-container
         *ngTemplateOutlet="popoverInner; context: { target: _sliderValueTargets.SINGLE_SLIDER, position: _position }"
     ></ng-container>
-</ng-template>
+</fd-popover>
 
-<ng-template #rangeSliderPopover1>
+<fd-popover
+    #rangeSliderPopover1
+    *ngIf="_isRange"
+    [triggers]="[]"
+    [closeOnOutsideClick]="false"
+    [appendTo]="rangeHandle1"
+    [disabled]="!tooltipMode"
+    placement="top"
+>
     <ng-container
         *ngTemplateOutlet="
             popoverInner;
             context: { target: _sliderValueTargets.RANGE_SLIDER1, position: _handle1Value }
         "
     ></ng-container>
-</ng-template>
+</fd-popover>
 
-<ng-template #rangeSliderPopover2>
+<fd-popover
+    #rangeSliderPopover2
+    *ngIf="_isRange"
+    [triggers]="[]"
+    [closeOnOutsideClick]="false"
+    [appendTo]="rangeHandle2"
+    [disabled]="!tooltipMode"
+    placement="top"
+>
     <ng-container
         *ngTemplateOutlet="
             popoverInner;
             context: { target: _sliderValueTargets.RANGE_SLIDER2, position: _handle2Value }
         "
     ></ng-container>
-</ng-template>
+</fd-popover>
 
 <ng-template #popoverInner let-target="target" let-position="position">
     <div class="fd-slider--tooltip-wrapper" (click)="$event.stopPropagation()" (keydown)="$event.stopPropagation()">
